+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s0/s3/s4/s_0034_test.kt
@@ -1,16 +1,16 @@
 package s0.s0.s3.s4
 
-import utils.Test_
+import utils.T
 import kotlin.test.Test
 import s0.s0.s3.s4.s_0034_00.Solution as s_0034_00
 
 class s_0034_test {
     @Test
     fun test() {
-        val test = Test_(
+        val t = T(
             this::class,
             s_0034_00::class,
         )
-        test.run()
+        t.run()
     }
 }