+++ workspace/solutions/kotlin/lib/src/test/kotlin/s0/s3/s0/s3/s_0303_test.kt
@@ -1,15 +1,15 @@
-package s0.s9.s1.s2
+package s0.s3.s0.s3
 
 import utils.T
 import kotlin.test.Test
-import s0.s9.s1.s2.s_0912_00.Solution as s_0912_00
+import s0.s3.s0.s3.s_0303_00.NumArray as s_0303_00
 
-class s_0912_test {
+class s_0303_test {
     @Test
     fun test() {
         val t = T(
             this::class,
-            s_0912_00::class,
+            s_0303_00::class,
         )
         t.run()
     }