+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - [303. 区域和检索 - 数组不可变 - 力扣（LeetCode）](https://leetcode.cn/problems/range-sum-query-immutable/description/)
 - 改进 `go` 题解的自动化测试 | Import automated tests for `go` solutions
 - 改进 `java` 题解的自动化测试 | Import automated tests for `java` solutions
+- 改进 `kotlin` 题解的自动化测试 | Import automated tests for `kotlin` solutions
 
 ## 2024-08-24